utils: Move TextListWidget to utils/graphicstextlist.py
